LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-fetch-error.sub-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200

PASS import() must reject when there is a wrong MIME type
PASS import() must reject with a different error object for each import when there is a wrong MIME type
PASS import() must reject when there is a wrong MIME type of a dependency
PASS import() must reject with a different error object for each import when there is a wrong MIME type of a dependency
PASS import() must reject when there is a 404
PASS import() must reject with a different error object for each import when there is a 404
PASS import() must reject when there is a 404 of a dependency
PASS import() must reject with a different error object for each import when there is a 404 of a dependency
PASS import() must reject when there is a 500
PASS import() must reject with a different error object for each import when there is a 500
PASS import() must reject when there is a 500 of a dependency
PASS import() must reject with a different error object for each import when there is a 500 of a dependency
PASS import() must reject when there is a cross-origin module (without CORS)
PASS import() must reject with a different error object for each import when there is a cross-origin module (without CORS)
PASS import() must reject when there is a cross-origin module dependency (without CORS)
PASS import() must reject with a different error object for each import when there is a cross-origin module dependency (without CORS)
PASS import() fetch errors must be cached
